Final Student Database with Templated HTML
This is the final version of the web app GUI for a student database. It is just like the other one, but this one now has a working lay out. Basically, I downloaded a template and used it as a base for my html.  I did changes like how to manipulate date from a postgresql thru psycopg2 and then passing it to the html using flask.